refactor: selectors, skip tokens
